Adding attachment mechanism
In order to keep the initial part of the report clean, an attachment
mechanism is added. Tests can submit larger pieces of text which
are displayed at the end while showing a summary message in the
initial part. Proof-of-concept for rpmlint and MD5-sum checks.
Removed the nowrap support, the attachment mechanism covers it's
actual usecases.