Unpack using rpm2cpio instead of rpm -i
Using rpm -i to install the srpm is to invasive. It stores
things in user's rpmbuild tree, and even has an option to
wipe it. For users without personal rpmbuild environment, this
might end up trying to write to system's rpmbuild tree.

This patch uses rpm2cpio to unpack the srpm to a local directory
instead.