Remove configuration file
The configuration file is not documented, and is also created
automatically causing some options to be sticky after first run.
This add lot of confusion. Removing for now, leaving hooks ion
settings .py for re-adding it.

Thoughts: most options are not really global, but rather per
report i. e., the need is to run f-r with the same options
as last time for a specific bug.