Bugfixes: handle build errors better, handle broken symlinks
Old code did not detect build errors, the shell pipe returns
code from last component i . e., grep which always is 0.
subcommand is not usable here, mock displays it output in  a
new shell and is not in Popen.stdout. So: dig for error
messages in build.log to understand what happened.

mock --rebuild --no-cleanup-after leaves broken symlinks in build
directory which needs to be deleted.