Introducing EXTRA test type
With the strong emphasis on guidelines, there is still need for some tests
which are just common sense, but not specified. Examples here: that
the spec in the URL matches the spec in the srpm, and running rpmlint
on installed package. The motivations are different, but it boils
down to the a new category.

I would prefer a name which was sorted after 'SHOULD', but me English
is too bad :(