Fix so-files in non-devel package check.
Method comment is OK, stating that so-files must be in a -devel
package if they are symlinks to versioned so-files. However,
the implemenatation just tests if there are so-files in non-devel
packages(s), returning False if so. This is plain wrong, so-files
in non-devel packages are perfectly OK in many situations.

New test returns True if all so-files in devel-packages. Else, gives
a 'fix or explain' message for the so-files and return
'inconclusive'.