Use local source when no upstream (#48)
For situations with git or similar releases, the Source0 is typically
just a filename, not an url. Comments in the file is there how to
create this file. This patch looks for a local file with this name
and uses it if it exists. This way, a reviewer can create the source
and then have f-r to check md5sum etc.