Remove bugzilla options assign, login, comment (#67)
These functions doesn't really fit in fedora-review's overall mission
to create a review template. Furthermore, the dependency python-bugzilla
nowadays contains the 'bugzilla' tool which can make a plethora of
actions to create and modify bugzilla bugs. Removing functions in order
to clean-up the CLI interface and keep the application as small as
possible. Adding a SEE ALSO bugzilla reference in the manpage.