CheckLicensInDoc: use rpm -qldp instead of %doc (issue #91).
Old code compares license files in source with contents in %doc
and %_docdir. This is unreliable e. g.,  when a directory is listed
there is no way to find out the files included.

New code compares license files with output from rpm -qldp which gives
the complete list of files and thus works also when dirs are listed.