Do not try to split %config lines
We assumed there was a space between config and place but this is not
neccesarily true. For example git.spec has:

%config(noreplace)%{_sysconfdir}/xinetd.d/git

So instead search the whole line for occurences. We should probably
work with expanded macros here, or do some other magic as
well...searching for datadir is probably not enough.