Add current settings.py directory on sys.path
This makes it possible to use plugins symlinked into python site-packages
directory and not clash with any git clones (i.e we'll always prefer git things
if running from git)

By default site-packages/ is on sys.path, but site-packages/FedoraReview is not.
So site-packages/FedoraReview/plugins would not be found by python-straight
plugin. With this change site-packages/FedoraReview gets added on sys.path
temporarily to solve this