Fixing printed 'Buildroot: None'
Current code displays 'Buildroot: None' in last part of report on
occasions. This is really about an import path problem, the Mock
instance used in checks is actually not the same as in other parts.
Fixing this in release branch means too big changes (shrinking the
module interface and overall import paths fixes). Making temporary
fix with new property.