Move is_plugin_installed into Checks class
Since we changed the approach a bit, registry is only concerned with itself.
However we need to ask inside basic checks if there is a registry for given group
with external_plugin set to True. That way we can create stub checks that will
output warnings about missing external plugins if needed.