Adjust license checker to take into account new-style %license files
If fedora-review is called with --define EPEL5 or --define EPEL6, and
a license file exists in the sources, CheckLicensInDoc checks if this file
is among the files marked as %doc as before.

Otherwise, it checks if the file is marked as %license, and fails if either
the file is not marked as %license or it is marked as %doc.

Signed-off-by: Michel Alexandre Salim <salimma@fedoraproject.org>