Fix false positive for Packager: tag (bz #1146442).
Logic is that if macro expands and we can't find the corresponding
header re in the spec, them that's a false positive we can drop.
Will still generate a false positive e. g., if there is a Packager
macro in ~/.rpmmacros and a line starting with packager: in a text
block such as the description. Let's live with that.